examate_project/requirement.txt
asgiref==3.7.2
chardet==5.2.0
coverage==7.3.2
Django==4.2.6
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
mysqlclient==2.2.0
packaging==23.2
pillow==10.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2023.3.post1
reportlab==4.0.9
sqlparse==0.4.4
typing_extensions==4.8.0
xhtml2pdf==0.2.15
google-api-python-client==2.124.0
fcm-django==2.0.1
firebase-admin==6.5.0
pyfcm==1.5.4
stripe==9.3.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2




